some fix in admin

+++ admin/items/categorias/delete.php
@@ -42,25 +42,15 @@ if(!$categoria){
 try {
     #Buscar la categoria en ingles para eliminarla
     $categoriaEn = CategoriaEng::where("id_categoria", $categoria->id);
-    $categoriaEn->delete();
+    if($categoriaEn){
+        $backupCategoriaEn = CategoriaEng::where("id_categoria", $categoria->id);
+        $categoriaEn->delete();
+    }
 
     #Eliminar la categoria en español
     $categoria->delete();
 
-
-    if($_SESSION["lang"] == "es"){
-        $_SESSION["alert"]["message"] = "La categoria ha sido borrada exitosamente!";
-        $_SESSION["alert"]["type"] = "success";
-        $_SESSION["alert"]["alert"] = "simple";
-    }
-    else{
-        $_SESSION["alert"]["message"] = "The category has been deleted successfully!";
-        $_SESSION["alert"]["type"] = "success";
-        $_SESSION["alert"]["alert"] = "simple";
-    }
-    header("location: /admin/items/categorias");
-
-} catch (\Throwable $th) {
+} catch (mysqli_sql_exception) {
     if($_SESSION["lang"] == "es"){
         $_SESSION["alert"]["message"] = "Hubo un error, trata de borrar otra cosa para intentar eliminar este item!";
         $_SESSION["alert"]["type"] = "error";
@@ -69,6 +59,26 @@ try {
     else{
         $_SESSION["alert"]["message"] = "There was a mistake, try to erase something else to try to eliminate this item!";
         $_SESSION["alert"]["type"] = "error";
+        $_SESSION["alert"]["alert"] = "simple";
+    }
+    if($backupCategoriaEn){
+        $backupCategoriaEn->setData("id", '');
+        $backupCategoriaEn->save();
+    }
+
+    header("location: /admin/items/categorias");
+    $error = true;
+}
+
+if(!$error){
+    if($_SESSION["lang"] == "es"){
+        $_SESSION["alert"]["message"] = "La categoria ha sido borrada exitosamente!";
+        $_SESSION["alert"]["type"] = "success";
+        $_SESSION["alert"]["alert"] = "simple";
+    }
+    else{
+        $_SESSION["alert"]["message"] = "The category has been deleted successfully!";
+        $_SESSION["alert"]["type"] = "success";
         $_SESSION["alert"]["alert"] = "simple";
     }
     header("location: /admin/items/categorias");
